Move painters to the proper file
